Use native cryptography lib if installed
